+++ .gitignore
@@ -1,0 +1,4 @@
+.idea
+.vscode
+USPS
+*.pt
